requirements.txt
joblib==1.5.0
numpy==2.3.1
pandas==2.3.0
pdfminer_six==20250327
PyPDF2==3.0.1
scikit_learn==1.7.0
SQLAlchemy==2.0.41
streamlit==1.45.1
plotly==5.24.0
lightgbm==4.5.0
psycopg2-binary==2.9.10
openpyxl==3.1.5
reportlab==4.2.5